requirements.txt
django
djangorestframework
Pillow
web3
python-dotenv